# Conflicts:
#	app/src/main/res/layout/activity_movie_detail.xml
#	app/src/main/res/layout/fragment_scan_qr_code.xml

+++ app/src/main/java/com/example/kotlinv2/model/Movie.kt
@@ -11,10 +11,10 @@ data class Movie(
     var original_language: String? = "",
     var title: String? = "",
     var backdrop_path: String? = "",
-    var popularity: Double? = 0.0,
+    var popularity: Float? = 0F,
     var vote_count: Int? = 0,
     var video: Boolean? = false,
-    var vote_average: Double? = 0.0
+    var vote_average: Float? = 0F
 ) {
     override fun toString(): String {
         return "PopularMovie(poster_path=$poster_path, adult=$adult, overview=$overview, release_date=$release_date, genre_ids=$genre_ids, id=$id, original_title=$original_title, original_language=$original_language, title=$title, backdrop_path=$backdrop_path, popularity=$popularity, vote_count=$vote_count, video=$video, vote_average=$vote_average)"
